update my markets to use filterbox
